feat: 🎸 add day 12, part A and B
part A is a BOGOSORT kind of solution, we just fill all the ? in with a
randomized value and see if it's correct. It does a million iterations
to arrive at (hopefully) the right solution, and could still be off. I
had to run it 10x just to make sure iI got the right value. T Part B
would be impossible to brute force, so we'd have to come up with a
better solution. I'm simply too dumb to do this though, and therefore
stole someone elses solution. I forgot who it was but thank you to that
person