feat: 🎸 add day 14, part A and B
Part B wants you to run rotations 1 billion times, for funsies. This
would take a long long time. 2 million items run on my machine in about
3 minutes, s so 1 billion items would take about 25 hours. Fortunately,
since this is a loop, we can just run 1000 iterations an dstill get the
right answer! Thanks to the people on reddit for discovering this. I
probably could've figured it out myself but I'm just not not built like
that. So Other than that, this day was pretty easy, I really enjoyed
part 1. Hope you have an amazing day! <3