feat: 🎸 add day 18, part A and B
This one was not that difficult, but I did learn about the shoelace
formula. That was pretty interesting, getting that one to work. I also
had lots of fun with CRLF and LF (LF wins always btw, microsoft pls
fix). In CRLF my parsing didn't work, and so it would spit out the wrong
answer. We love to see it. Anyway, it's all good now. I hope you had a
great day, keep up the great work!