feat: 🎸 add day 20, part A and B
This was another really fun one. Part one was a breeze, but really fun
to implement with a queue like that. Part 2 was harder, and I had to
pull out the LCM again. I'm still not smart enough to detect these
patterns by myself, so I once again had to reach for hyperneutrino's
video and translate the Python code to TypeScript. I was almost there
though! I hope tomorrow's is just as fun, and I hope you have a great
day!