feat: 🎸 add day 24, part A and B
This one was impossible for me. It was super math heavy, all of which I
had never done or even seen before. I had a lot of trouble. For part 1 I
could follow HyperNeutrino's video, but for part 2 they used a Python
library, which we obviously don't have available in JavaScript.
Therefore I had to look for a different implementation to follow, and
ayoubzulfiqar came to the rescue with their implementation of the
problem. I wish I could've solved this myself, but at least I understand
on a high level what happens and what's going on in the code. Mentally I
know on a high level how to solve the problem, but I'm not sure about
the specifics. It's still Christmas today, so I hope you have an amazing
day and MERRY CHRISTMAS 🎄🎄🎄